recruiting_analytics_backend/requirements/base.txt
django
urllib3
attrs
cattrs
gunicorn
django-countries
django-nested-admin
django-grappelli
requests
django-cors-headers
djangorestframework
djangorestframework_jwt==1.11.0
django-environ
